init copy multi initiative
